Pluginable registry client.
