+++ etc/conf/app.conf
@@ -14,9 +14,6 @@ httpport = 30100
 # 'etcd' means app running as an etcd agent
 # 'embeded_etcd' means app running as an etcd server
 registry_plugin = etcd
-# Indicates the time in seconds that service center waits since it receive
-# the last heartbeat before it remove the instance from registry service
-registry_lease_duration = 30s
 
 # registry address
 # registry_plugin equals to 'embeded_etcd', example:
