+++ server/core/registry/store/indexer.go
@@ -16,7 +16,7 @@ package store
 import (
 	"github.com/ServiceComb/service-center/pkg/util"
 	pb "github.com/ServiceComb/service-center/server/core/proto"
-	"github.com/ServiceComb/service-center/server/core/registry"
+	"github.com/ServiceComb/service-center/server/infra/registry"
 	"github.com/coreos/etcd/mvcc/mvccpb"
 	"golang.org/x/net/context"
 	"strings"
@@ -60,7 +60,7 @@ func (i *Indexer) Search(ctx context.Context, opts ...registry.PluginOpOption) (
 		(op.Offset >= 0 && op.Limit > 0) {
 		util.Logger().Debugf("search %s match special options, request etcd server, opts: %s",
 			i.cacheType, op)
-		return registry.GetRegisterCenter().Do(ctx, opts...)
+		return backend.GetRegisterCenter().Do(ctx, opts...)
 	}
 
 	if op.Prefix {
@@ -75,7 +75,7 @@ func (i *Indexer) Search(ctx context.Context, opts ...registry.PluginOpOption) (
 
 		util.Logger().Debugf("can not find any key from %s cache with prefix, request etcd server, key: %s",
 			i.cacheType, key)
-		return registry.GetRegisterCenter().Do(ctx, opts...)
+		return backend.GetRegisterCenter().Do(ctx, opts...)
 	}
 
 	resp := &registry.PluginResponse{
@@ -95,7 +95,7 @@ func (i *Indexer) Search(ctx context.Context, opts ...registry.PluginOpOption) (
 		}
 
 		util.Logger().Debugf("%s cache does not store this key, request etcd server, key: %s", i.cacheType, key)
-		return registry.GetRegisterCenter().Do(ctx, opts...)
+		return backend.GetRegisterCenter().Do(ctx, opts...)
 	}
 
 	cacheData := i.Cache().Data(key)
@@ -106,7 +106,7 @@ func (i *Indexer) Search(ctx context.Context, opts ...registry.PluginOpOption) (
 
 		util.Logger().Debugf("do not match any key in %s cache store, request etcd server, key: %s",
 			i.cacheType, key)
-		return registry.GetRegisterCenter().Do(ctx, opts...)
+		return backend.GetRegisterCenter().Do(ctx, opts...)
 	}
 
 	resp.Count = 1
