+++ server/plugin/infra/quota/buildin/buildin.go
@@ -17,9 +17,9 @@ import (
 	"fmt"
 	"github.com/ServiceComb/service-center/pkg/util"
 	"github.com/ServiceComb/service-center/server/core"
-	"github.com/ServiceComb/service-center/server/core/registry"
 	"github.com/ServiceComb/service-center/server/core/registry/store"
 	"github.com/ServiceComb/service-center/server/infra/quota"
+	"github.com/ServiceComb/service-center/server/infra/registry"
 	mgr "github.com/ServiceComb/service-center/server/plugin"
 	"golang.org/x/net/context"
 )
